Refactor feedItems endpoint to put logic in controller.
